Isolate tests into separate files
