+++ firmware/projects/Demo/TestCangen/test/cantest.hpp
@@ -1,0 +1,40 @@
+#include "../generated/can/test_rx_bus.hpp"
+#include "gtest/gtest.h"
+#include "shared/periph/can.hpp"
+
+#define EXPECT_MSG_EQ(x, y)                                      \
+    ASSERT_EQ((x).data_length, (y).data_length);                 \
+    EXPECT_EQ((x).id, (y).id);                                   \
+    EXPECT_EQ((x).is_extended_frame, (y).is_extended_frame);     \
+    for (size_t i = 0; i < (x).data_length; ++i) {               \
+        EXPECT_EQ((x).data[i], (y).data[i]) << "at index " << i; \
+    }
+
+using namespace generated::can;
+using shared::can::RawMessage;
+
+class Bus : public testing::Test {
+    class DummyCanBase : public shared::periph::CanBase {
+    public:
+        void Send(const shared::can::RawMessage& m) override {
+            throw std::runtime_error(
+                "Dummy bus is only for receiving, not sending.");
+        }
+
+    private:
+        uint32_t GetTimestamp() const override {
+            return 0;
+        }
+    };
+
+protected:
+    Bus() : base(), bus(base) {}
+
+    DummyCanBase base;
+    Test_rxBus bus;
+};
+
+template <typename T>
+class CanTest : public Bus,
+                public testing::WithParamInterface<std::tuple<RawMessage, T>> {
+};
